src/prompts.py
"""
System Prompts for the Medical Chatbot.
Designed to be friendly, professional, and prevent hallucinations.
"""

from langchain_core.prompts import PromptTemplate, ChatPromptTemplate
from config import MEDICAL_DISCLAIMER, SEVERITY_KEYWORDS


def get_system_prompt() -> str:
    """
    Get the main system prompt for the medical chatbot.
    
    Returns:
        System prompt string
    """
    return """You are a professional doctor conducting a medical consultation. Engage in a natural, conversational dialogue.

## Guidelines:
- Be direct and professional
- Suggest specific medications and treatment regimens from the knowledge base
- Ask relevant follow-up questions to better understand the patient's condition
- Keep responses concise and helpful
- Do NOT use headers, numbered steps, or bullet points

## Knowledge Base:
{context}

## Patient Question:
{question}

## Your Response:
1. Address the patient's current concern with specific recommendations
2. Ask ONE relevant follow-up question to gather more information
3. End with a brief disclaimer

⚠️ **DISCLAIMER**: This is informational only. See a doctor for medical advice."""


def get_medical_prompt_template() -> PromptTemplate:
    """
    Get the prompt template for medical queries.
    
    Returns:
        PromptTemplate for medical questions
    """
    template = get_system_prompt()
    
    return PromptTemplate(
        input_variables=["conversation_history", "question", "context"],
        template=template
    )


def get_medical_chat_prompt() -> ChatPromptTemplate:
    """
    Get the chat prompt template for conversational interface.
    
    Returns:
        ChatPromptTemplate for medical chat
    """
    return ChatPromptTemplate.from_messages([
        ("system", get_system_prompt()),
        ("placeholder", "{conversation_history}"),
        ("human", "{question}")
    ])


def get_followup_prompt() -> str:
    """
    Get prompt for follow-up questions.
    
    Returns:
        Follow-up prompt string
    """
    return """This is a follow-up question in an ongoing medical conversation.

Previous conversation context:
{conversation_history}

The user is asking a follow-up question. 
Please refer to the previous context when answering.
If this question seems unrelated to the previous conversation, acknowledge that politely.

Current question: {question}

Remember:
- Reference previous symptoms or concerns if relevant
- Provide consistent, helpful responses
- Always include the medical disclaimer
"""


def get_severity_awareness_prompt() -> str:
    """
    Get prompt for detecting severe symptoms.
    
    Returns:
        Severity awareness prompt
    """
    keywords = ", ".join(SEVERITY_KEYWORDS)
    
    return f"""Analyze the following symptoms for severity.

WARNING KEYWORDS: {keywords}

User symptoms: {{symptoms}}

If any WARNING KEYWORDS are present OR symptoms appear severe:
- Immediately advise seeking emergency medical attention
- Provide clear, urgent language
- Do NOT attempt to provide detailed information - prioritize getting them help

If symptoms are mild/moderate:
- Provide helpful information from the knowledge base
- Suggest when to see a doctor if symptoms worsen
- Include self-care recommendations

Response:"""


def get_source_citation_prompt() -> str:
    """
    Get prompt for source citation.
    
    Returns:
        Source citation prompt
    """
    return """Based on the following medical information, provide an answer to the user's question.

Medical Information:
{{context}}

User Question:
{{question}}

Requirements:
- Base your answer ONLY on the provided information
- If the information doesn't fully answer the question, acknowledge this
- Cite the source(s) when possible
- Include the medical disclaimer
- Keep it concise and helpful

Answer:"""


def get_condensed_history_prompt() -> str:
    """
    Get prompt for condensing conversation history.
    
    Returns:
        Condensed history prompt
    """
    return """The following is a conversation history. Condense it to the most important information for understanding the user's medical situation.

Conversation:
{conversation}

Provide a brief summary (2-3 sentences) of the key medical context:"""
