requirements.txt
# Medical Chatbot Dependencies
# ============================

# Core Dependencies
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0

# Embeddings and Vector Store
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
langchain-huggingface>=0.1.0

# Document Loaders
pypdf>=4.0.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
langchain-experimental>=0.0.0

# Configuration
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0

# Optional: For enhanced URL loading
# playwright>=1.40.0
# playwright install
